backend: Release accelerometer sensor when panel is powered off via DPMS
When the screen of a laptop or phone is turned off, there's no reason to
listen to accelerometer events and react to orientation changes, we can do
that just as well once the screen is turned on again. So add another point
where we inhibit the orientation manager on power_save_mode changes of the
monitor.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>