tests/sensors-proxy-mock: Don't wait for dbus client after creating the mock
We'll change mutter to no longer immediately Claim the accelerometer if
there's no reason to listen to rotation changes, so remove the initial wait
for a client to appear in the iio-sensor-proxy mock, as that will no longer
work in all cases.

With the next commit we'll introduce a new API for the SensorsProxyMock
where this wait can be selectively applied instead.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>