orientation-manager: Emit new signal ::sensor-active after claiming sensor
The ::orientation-changed signal won't emit when the sensor becomes active
but the new orientation is the same as the one that we have stored. To still
notify listeners that the sensor has been claimed and that they can now read
orientation via `meta_orientation_manager_get_orientation()`, emit a new
signal ::sensor-active.

Use this signal in the monitor manager for exactly the purpose mentioned
above.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>