tests/sensors-proxy-mock: Properly keep track of claiming clients
In tests, Mutter can claim a client two or more times at the same time,
using different DBus proxies. The orientation-manager-unit-tests create
their own MetaOrientationManager instance, which Claim()'s the sensor on top
of Mutters own MetaOrientationManager. This will lead to test failures once
mutter starts releasing the sensor in its own orientation manager when it's
unneeded.

To fix this, the mock has to handle the same client claiming twice, which it
currently doesn't. We do that by adding a claim-count to the internal struct
that keeps track of the clients which have claimed our sensor.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4119>