wayland: Simplify key focus handling at MetaWaylandInput
Tablet pads share focus with the keyboard, and will be already switched
by the MetaWaylandSeat when handling the keyboard focus. And in fact,
other things do also share the key focus (e.g. IMs), and this is not
really tied to wl_keyboard availability.

Later commits will make it more evident that we are dealing with the
one and only ClutterKeyFocus here, but let's not wait for that to remove
this.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4458>