backends: Make device implicit for MetaPointerConstraintImpl
From the name it is already clear it affects the mouse pointer, and
it implicitly warps the pointer cursor while enforcing the constraint.
Make the pointer queried also implicitly be the pointer cursor from
the passed seat, and avoid passing the device when enabling the
constraint.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4458>