clutter/color-state-params: Use extended sRGB for blending by default
This makes blending work as expected for sRGB content with HDR on,
fixing visual issues reported in the issue below and on
https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4387.

In turn, it introduces similar issues for linear content with HDR on.
https://gitlab.gnome.org/GNOME/mutter/-/issues/4084 describes a
longer-term plan for making this (and blending of content using
arbitrary other color states) work as expected.

Meanwhile, this is a trade-off which makes the behaviour consistent with
HDR on & off, and reflects the current reality that the vast majority of
content (100% of gnome-shell UI & Xwayland content, plus the vast
majority of native Wayland client content) is sRGB.

As a bonus, it fixes SW cursors being brighter than reference white with
HDR on, and the gnome-shell Looking Glass colour overlays.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4028
Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8264
Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3965

v2:
* Use CLUTTER*COLORSPACE* for blending_colorimetry. (Joan Torres)
v3:
* Use output max luminance with ref_is_1_0 for the blending color state
  to prevent spurious tone mapping.
* Bump COLOR_TRANSFORM_EPSILON to 0.05 in the color-state-transform
  test.
v4:
* Use output ref luminance as well.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4387>