wayland: Drop second accounting of keyboard modifiers
Nowadays, we get pressed/latched/locked modifiers in a detailed manner
from ClutterEvent itself, and we do not need to keep a second accounting
of these modifiers. This includes merging accessibility modifiers, and
special handling of numlock.

Drop this code, and rely entirely on modifiers as they arrive through
ClutterEvents, these will already have the correct a11y state and whatnot.
So far we cannot query the pressed keys for wl_keyboard.enter, so at the
moment preserve the xkb_state for this.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3340
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4546>