clutter/color-state-params: Add luminances_equal and needs_tone_mapping
This functions already exist giving color_state_params as args, now
add the same ones but giving luminances as args instead.

This will be used by the next commit when dealing with luminance mapping and
tone mapping on color_states that are not color_state_params, so we can only
rely on default luminances.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4102>