window/wayland: Update unconstrained rect on synchronous moves
The unconstrained rect is used to keep track of the position and size of
a window as if it wasn't constrained (e.g. maximized). It's only done in
some cases for Wayland due to the asyncronous nature of window resizes,
but to get up to date unconstrained rect when placing windows that
already has content in the floating state, also update the unconstrained
rect when a non-ack:ed move/resize call can move immediately.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4582>