wayland/xdg-shell: Pass initial config through move/resize machinery
This means configuration set by the configure signal handler (e.g.
fullscreen/maximized etc) gets properly handled by the relevant
constraints handling setting up the correct size.

Prior to this, the application of the initial commit was mistakenly
interpreted as a "client resize", which is meant to try to distinguish
between sporadic client resizes, while in this case, it was not. This
meant the unconstrained rect was not updated after placement, so make
that explicit. We can't make such a decision in the Wayland move/resize
implementation, since it doesn't have the necessary information.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4219
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4582>