wayland/xdg-session-state: Configure client surface when restoring
When restoring, we'd always resize according to the saved rectangle, but
with some special flags that effectively made it just move the logical
window position. This was done unconditionally to make sure a
maximized/fullscreen window was restored on the right monitor, but for
floating windows, the flags caused the client surface to not be
configured.

To improve this, first replace the flags with 'FORCE_MOVE', since that
is more aligned with what is expected to happen and has more sane side
effects, then only force-move prior to applying any of the tiled
(maximized/fullscreen/...) state. This means floating windows gets
properly configured via the move_resize() call, and non-floating windows
gets configured via their corresponding enter-tiled-state calls, and not
during the force-move call.

This will help with further cleanups to Wayland window configuration
logic.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4582>