wayland/surface: Don't treat buffer-less viewport as having a size
They effectively doesn't have one, since there is nothing that can be
stretched to the intended destination size. This fixes an issue where Qt
commits a surface state with a viewport before assigning the surface a
toplevel role. This caused the geometry to be incorrectly calculated to
the destination rectangle of the viewport, which caused confusion about
geometry changes and window resizes, when there was no such thing.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4582>