window: Notify about being created after being created
There are things that rely on the window being set on its owner (e.g.
the Wayland surface role), for example the function to get the toplevel
window of a Wayland surface. To let anything happening in
`MetaDisplay::window-created` work with the assumption that the window
is fully accounted for notify said signal after it was added to any
corresponding accounting structure.

This fixes the following crash:

    #0  meta_window_update_monitor (window=0x0, ...) at ../src/core/window.c:4132
    #1  meta_window_wayland_update_main_monitor (...) at ../src/wayland/meta-window-wayland.c:588
    #2  meta_window_update_monitor (...) at ../src/core/window.c:4138
    #3  meta_window_move_resize_internal (...) at ../src/core/window.c:4352
    #4  meta_window_move_resize (...) at ../src/core/window.c:4380
    #5  meta_window_set_maximize_flags (...) at ../src/core/window.c:2942
    #6  meta_window_maximize (...) at ../src/core/window.c:2953
    #7  on_window_created (...) at ../src/tests/window-created-tests.c:69
    ...

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4252
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4583>