window/wayland: Store a reference the the wayland client
We now have a MetaWaylandClient that exists for all clients, which can
even hang around for longer if we take a reference. So start using it in
MetaWindowWayland so we always know about the client, even if the client
disconnected or the window resource was destroyed.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4585>