wayland/client: Store the pid of the client
If one queries the wayland connection fd of clients created with
meta_wayland_client_new_subprocess or meta_wayland_client_new_create,
about the pid of the process which created it, it will correctly return
that the process mutter runs in has created it.

This isn't really useful though. We want to know who actually is the
primary user of the connection.

In the meta_wayland_client_new_subprocess case, it is the subprocess
pid.

In the meta_wayland_client_new_create case, we have to ask the caller.
The service channel uses it, and there the user of the dbus service is
the primary user.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4585>