clutter/gesture: Filter out some unhandled events manually
ClutterGesture is currently only supposed to handle pointer and touch
events, everything else is not supported.

It seems that some touchpad gesture events are making it through the
`get_sequence_data()` check, which should find the ClutterGesture internal
sequence data for any events that we're tracking (and filter out any events
that we aren't tracking). The reason for that is that pointer events with
buttons pressed down come without an event sequence and
`get_sequence_data()` tries to map those to ClutterGestures internal
sequence tracker anyway.

The touchpad gesture (and also scroll) events looks similar enough (ie. they
come from the same input device and also have no event sequence) that they
fall through the cracks and `get_sequence_data()` finds a seq_data entry for
them. This then causes assertion failures in the event handling code that
follows.

To deal with this (at least until all of the events that we want to handle
come with proper ClutterEventSequences), simply filter out those event types
that we aren't supposed to handle.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4590>