window/wayland: Keep and use window drag hen config is acked
When we're moving a maximized window using a window drag, it will after
a threshold snap out of its maximized state. When a unmaximize
configuration is pending, don't actually move the window, but amend the
stored position in the pending configuration according to the window
drag position calculations.

Also keep track of what drag caused what unmaximize request, so that
when the unmaximized surface configuration is acknowledged, an up to
date position derived from the window drag can be calculated. Keep a
reference to the drag, so that even if the drag ended a position can be
calculated from the last state of the drag.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4281
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4607>