eis-client: Clear keyboard signal handlers when device removed
If an EIS client releases a keyboard device but doesn't unbind the seat
capability, we'd still listen on keyboard related signals that assumed
the keyboard was still there. Fix this by disconnecting the signal
handlers also when the device is released.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4632>