clutter: Fix coordinates in crossing events
When making a sprite update itself based on new device coordinates,
we are updating the actor (and generating crossing events) first,
then updating the sprite position. This results in the crossing
events missing the up-to-date device coordinates.

The right order is the opposite, we should first update the sprite
coordinates, then update the actor and generate the crossing events
with the right coordinates.

This fixes crossing events having an out of date position, possibly
0,0 for touchpoints that just landed on the screen. This was
noticeable in window dragging, as it would use this crossing event
position to figure out the anchor point.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4245
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4640>