compositor: Don't assume there's an active ClutterFrame during X11 paint
The paint_context created in `create_framebuffer_from_window_actor` is
too short lived to have a `ClutterFrame` associated with it. That only
exists during a view paint. Since we're clearly not synchronised when
this happens, don't suggest we know the frame counter.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4254
Fixes: b5ca2a46d52 ("compositor,x11: Use per-view frame counter ...")
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4612>