backends/native: Update keymap state before queuing up key event
This should make it guaranteed and race-free that a key event handler
in the main thread can query the keymap state and see its own effects
already applied.

Both things get queued up for the main thread to handle, and with the
previous order there seem to be slim chances that the main thread
context is already spun for the queued up event, not leaving changes
for the keymap update to be processed before that despite having
G_PRIORITY_HIGH.

Since MetaWaylandKeyboard relies on ClutterKeymap state being current
at the time to handle an event, this fixes ghost pressed modifiers
in clients when those slim chances are met.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4315
Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4266
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4653>