clutter: Bail out in gesture event handler on sprite-less events
Gestures mandate a pointing presence, thus require sprites for it to
work. However, this variable could silently be NULL in case of key
events and other events related to the keyboard focus, this could
still happen to go through depending on the gesture state (e.g.
interaction with the actual pointer), causing an assert.

Make it clear that the ClutterSprite getter for the given event
may return NULL, and make ClutterGesture bail out on it.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4340
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4676>