wayland: Use xdg_session_management_v1 from wayland-protocols
Move away from our xx prefixed private interface, and adapt to the
minor changes that happened during the upstreaming of this protocol.

In the interim, still use our own internal copy of the protocol if
a recent enough version of wayland-protocols is not found.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4806>